+++ protected/humhub/modules/content/widgets/views/labels.php
@@ -1,23 +1,28 @@
 <?php
+
+use humhub\modules\post\models\Post;
+
 /**
  * This view shows common labels for wall entries.
  * Its used by WallEntryLabelWidget.
  *
  * @property Mixed $object the content object (e.g. Post)
  *
- * @package humhub.modules_core.wall
  * @since 0.5
  */
 ?>
 <?php if ($object->content->isPinned()) : ?>
-    <span class="label label-danger"><?php echo Yii::t('ContentModule.widgets_views_label', 'Pinned'); ?></span>
+    <span class="label label-danger"><?= Yii::t('ContentModule.widgets_views_label', 'Pinned'); ?></span>
 <?php endif; ?>
 
 <?php if ($object->content->isArchived()) : ?>
-    <span class="label label-warning"><?php echo Yii::t('ContentModule.widgets_views_label', 'Archived'); ?></span>
+    <span class="label label-warning"><?= Yii::t('ContentModule.widgets_views_label', 'Archived'); ?></span>
 <?php endif; ?>
 
 <?php if ($object->content->isPublic()) : ?>
-    <span class="label label-success"><?php echo Yii::t('ContentModule.widgets_views_label', 'Public'); ?></span>
+    <span class="label label-success"><?= Yii::t('ContentModule.widgets_views_label', 'Public'); ?></span>
 <?php endif; ?>
 
+<?php if (!$object instanceof Post) : ?>
+    <span class="label label-default"><?= $object->getContentName(); ?></span>
+<?php endif; ?>