+++ protected/humhub/modules/content/widgets/views/wallEntry.php
@@ -1,101 +1,86 @@
 <?php
 
-use yii\helpers\Html;
-use humhub\modules\user\models\User;
+use humhub\libs\Html;
+use humhub\widgets\TimeAgo;
 use humhub\modules\space\models\Space;
-use humhub\modules\content\components\ContentContainerController;
-
-$user = $object->content->createdBy;
-$container = $object->content->container;
+use humhub\modules\user\widgets\Image as UserImage;
+use humhub\modules\content\widgets\WallEntryControls;
+use humhub\modules\space\widgets\Image as SpaceImage;
+use humhub\modules\content\widgets\WallEntryAddons;
+use humhub\modules\content\widgets\WallEntryLabels;
 ?>
 
-<div class="panel panel-default wall_<?php echo $object->getUniqueId(); ?>">
+<div class="panel panel-default wall_<?= $object->getUniqueId(); ?>">
     <div class="panel-body">
 
         <div class="media">
             <!-- since v1.2 -->
             <div class="stream-entry-loader"></div>
-            
+
             <!-- start: show wall entry options -->
             <ul class="nav nav-pills preferences">
                 <li class="dropdown ">
                     <a class="dropdown-toggle" data-toggle="dropdown" href="#"><i class="fa fa-angle-down"></i></a>
                     <ul class="dropdown-menu pull-right">
-                        <?php echo \humhub\modules\content\widgets\WallEntryControls::widget(['object' => $object, 'wallEntryWidget' => $wallEntryWidget]); ?>
+                        <?= WallEntryControls::widget(['object' => $object, 'wallEntryWidget' => $wallEntryWidget]); ?>
                     </ul>
                 </li>
             </ul>
             <!-- end: show wall entry options -->
 
-            <a href="<?php echo $user->getUrl(); ?>" class="pull-left">
-                <img class="media-object img-rounded user-image user-<?php echo $user->guid; ?>" alt="40x40"
-                     data-src="holder.js/40x40" style="width: 40px; height: 40px;"
-                     src="<?php echo $user->getProfileImage()->getUrl(); ?>"
-                     width="40" height="40"/>
-            </a>
+            <?=
+            UserImage::widget([
+                'user' => $user,
+                'width' => 40,
+                'htmlOptions' => ['class' => 'pull-left']
+            ]);
+            ?>
 
-            <!-- Show space image, if you are outside from a space -->
-            <?php if (!Yii::$app->controller instanceof ContentContainerController && $object->content->container instanceof Space): ?>
-                <?php echo \humhub\modules\space\widgets\Image::widget([
-                    'space' => $object->content->container,
+            <?php if ($showContentContainer && $container instanceof Space): ?>
+                <?=
+                SpaceImage::widget([
+                    'space' => $container,
                     'width' => 20,
-                    'htmlOptions' => [
-                        'class' => 'img-space',
-                    ],
+                    'htmlOptions' => ['class' => 'img-space'],
                     'link' => 'true',
-                    'linkOptions' => [
-                        'class' => 'pull-left',
-                    ],
-                ]); ?>
-
+                    'linkOptions' => ['class' => 'pull-left'],
+                ]);
+                ?>
             <?php endif; ?>
 
-
             <div class="media-body">
-
-                <!-- show username with link and creation time-->
-                <h4 class="media-heading"><a
-                        href="<?php echo $user->getUrl(); ?>"><?php echo Html::encode($user->displayName); ?></a>
-                    <small>
-
-                        <!-- show profile name -->
-                        <?php if (!Yii::$app->controller instanceof ContentContainerController && $container instanceof User && $container->id != $user->id): ?>
+                <div class="media-heading">
+                    <?= Html::containerLink($user); ?>
+                    <?php if ($showContentContainer): ?>
+                        <span class="viaLink">
                             <i class="fa fa-caret-right" aria-hidden="true"></i>
-                            <strong><a href="<?php echo $container->getUrl(); ?>"><?php echo Html::encode($container->displayName); ?></a></strong>&nbsp;
-                        <?php endif; ?>
+                            <?= Html::containerLink($container); ?>
+                        </span>
+                    <?php endif; ?>
 
-                        <?php echo \humhub\widgets\TimeAgo::widget(['timestamp' => $object->content->created_at]); ?>
-
-                        <?php if ($object->content->created_at !== $object->content->updated_at && $object->content->updated_at != ''): ?>
-                            (<?php echo Yii::t('ContentModule.views_wallLayout', 'Updated :timeago', array(':timeago' => \humhub\widgets\TimeAgo::widget(['timestamp' => $object->content->updated_at]))); ?>)
-                        <?php endif; ?>
-
-                        <!-- show space name -->
-                        <?php if (!Yii::$app->controller instanceof ContentContainerController && $container instanceof Space): ?>
-                            <?php echo Yii::t('ContentModule.views_wallLayout', 'in'); ?> <strong><a
-                                    href="<?php echo $container->getUrl(); ?>"><?php echo Html::encode($container->name); ?></a></strong>&nbsp;
-                        <?php endif; ?>
-
-                        <?php echo \humhub\modules\content\widgets\WallEntryLabels::widget(['object' => $object]); ?>
-
-                    </small>
-                </h4>
-                <h5><?php echo Html::encode($user->profile->title); ?></h5>
-
+                    <div class="pull-right labels">
+                        <?= WallEntryLabels::widget(['object' => $object]); ?>
+                    </div>
+                </div>
+                <div class="media-subheading">
+                    <?= TimeAgo::widget(['timestamp' => $createdAt]); ?>
+                    <?php if ($updatedAt !== null) : ?>
+                        &middot;
+                        <span class="tt" title="<?= Yii::$app->formatter->asDateTime($updatedAt); ?>">Updated</span>
+                    <?php endif; ?>
+                </div>
             </div>
             <hr/>
 
-            <div class="content" id="wall_content_<?php echo $object->getUniqueId(); ?>">
-                <?php if (!$object instanceof \humhub\modules\post\models\Post) : ?>
-                    <span class="label label-default pull-right"><?php echo $object->getContentName(); ?></span>
-                <?php endif; ?>
-                <?php echo $content; ?>
+            <div class="content" id="wall_content_<?= $object->getUniqueId(); ?>">
+                <?= $content; ?>
             </div>
-            
+
             <!-- wall-entry-addons class required since 1.2 -->
             <div class="stream-entry-addons clearfix">
-                <?php echo \humhub\modules\content\widgets\WallEntryAddons::widget(['object' => $object]); ?>
+                <?= WallEntryAddons::widget(['object' => $object]); ?>
             </div>
+
         </div>
     </div>
 </div>